+++ app/Http/Controllers/UserController.php
@@ -5,45 +5,72 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\User;
 use App\Http\Resources\UserResource;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Validator;
 
 class UserController extends Controller
 {
 
     public function index()
     {
-        return response(UserResource::collection(User::all(), 200));
+        if (Auth::user()['role_id']==1){
+            return response(UserResource::collection(User::all(), 200));
+        } else {
+            return response(['message'=>'No permission!'], 403);
+        }
     }
 
     public function store(Request $request)
     {
-        $user=$request->validate([
-            'email'=>'required',
-            'name'=>'required',
-            'password'=>'required'
-        ]);
-        return response(User::create($user), 201);
+        if (Auth::user()['role_id']==1){
+            return response(User::create([
+                'name' => $request->input('name'),
+                'email' => $request->input('email'),
+                'password' => $request->input('password'),
+                'role_id'=> $request->input('role_id')!=NULL
+                    ? $request->input('role_id')
+                    : 2,
+            ]), 201);
+        } else {
+            return response(['message'=>'No permission!'], 403);
+        }   
     }
+
 
     public function show(User $user)
     {
-        return response(new UserResource($user), 200);
+        if (Auth::user()['id']==$user['id'] || Auth::user()['role_id']==1 ){
+            return response(new UserResource($user), 200);
+        } else {
+            return response(['message'=>'No permission!'], 403);
+        }
     }
 
     public function update(Request $request, User $user)
     {
-        $data=$request->validate([
-            'email'=>'required',
-            'name'=>'required',
-            'password'=>'required'
+        if (Auth::user()['id']==$user['id'] || Auth::user()['role_id']==1 ){
+        $validate=Validator::make($request->toArray(),[
+                'name'=>'nullable',
+                'email'=>'nullable',
+                'password'=>'nullable'
         ]);
-        
-        $user->update($data);
-        return response($user, 201 );
+        if ($validate->fails()){
+            return response($validate->errors(), 400);
+        }
+        $user->update($validate->validate());
+            return response($user, 201);
+        } else {
+            return response(['message'=>'No permission!'], 403);
+        }
     }
 
     public function destroy(User $user)
     {
-        $user->delete();
-        return response(null, 204);
+        if (Auth::user()['id']==$user['id'] || Auth::user()['role_id']==1 ){
+            $user->delete();
+            return response(null, 204);
+        } else {
+            return response(['message'=>'No permission!'], 403);
+        }
     }
 }
